add functions for convert date and set data

+++ datebook/datebook/EventListModule/Model/EventListModel.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct EventListModel: Decodable {
     let id: Int
-    let dataStart: Int
-    let dataEnd: Int
+    let dataStart: Double
+    let dataEnd: Double
     let name: String
     let description: String
